chatbot-backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
aiosqlite==0.19.0
alembic==1.13.1

# Pydantic settings
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP client for AI APIs and WhatsApp
httpx==0.26.0

# Logging
python-json-logger==2.0.7

# Health check
requests==2.31.0